clean up shared code bumped engine version updated batcher settings cleaned up middleware code writes & queries added frontend envs made paima engine config global in frontend
