+++ templates/chess/packages/client/node/src/state-machine/persist/zombie.ts
@@ -18,7 +18,7 @@ export function scheduleZombieRound(
   block_height: number,
 ): SQLUpdate {
   const params: INewScheduledHeightDataParams = {
-    from_address: "0x0",
+    from_address: "0x0", // TODO This should be a Paima Engine Constant
     future_block_height: block_height,
     input_data: createZombieInput(lobbyId),
   };
