+++ templates/chess/packages/frontend/client/.env
@@ -1,0 +1,2 @@
+VITE_BATCHER_URL="http://localhost:3334"
+VITE_PAIMA_ENGINE_NODE_URL="http://localhost:9999"
